Вынесение изменяемых параметров в конфигурационный json файл

+++ KoboldTgBot/Errors/LLMEmptyAnswerException.cs
@@ -2,7 +2,7 @@
 {
     public sealed class LLMEmptyAnswerException : LLMException
     {
-        public LLMEmptyAnswerException(string propmpt, ushort maxLength, float temperature, float topPSampling, float repetitionPenalty) : base("Empty asnwer from LLM", propmpt, maxLength, temperature, topPSampling, repetitionPenalty)
+        public LLMEmptyAnswerException(string propmpt, int maxLength, float temperature, float topPSampling, float repetitionPenalty) : base("Empty asnwer from LLM", propmpt, maxLength, temperature, topPSampling, repetitionPenalty)
         {
         }
     }
